lazy but effectively named method i think.
helps make the code more readable by giving that hard to read ‘where’
statement a simple name to refer to it by.
i considered putting the .Split(‘,’) in there as well. but anytime you
end up putting the word ‘and’ in a method name, you are probably going
wrong somewhere.